refactor: entrance 구현하려다가 실패. spec 4끝나고 다시 도전할 예정

+++ src/main/java/com/nhnacademy/exam/parkingservice/ParkingLotPaymentService.java
@@ -11,8 +11,8 @@ public class ParkingLotPaymentService {
     ParkingFee parkingFee;
 
     public static final int TENMINUTE = 600;
-    public static final Long THIRTYMINUTE = 1800L;
-    public static final Long ONEHOUR = 3600L;
+    public static final int THIRTYMINUTE = 1800;
+    public static final int ONEHOUR = 3600;
     public static final int ONEDAY = 86400;
 
     public ParkingLotPaymentService(ParkingLotRepository parkingLotRepository,
